feat(data-formats): move StoredItem from rendezvous
The `StoredItem` struct is no longer needed just in the rendezvous server, we
are making it public and moving it to the data-formats lib so that it can be
easily included in other modules.
