+++ data-formats/src/serializable.rs
@@ -1,3 +1,6 @@
+use crate::cborparser::{ParsedArray, ParsedArrayBuilder, ParsedArraySize2};
+use crate::publickey::PublicKey;
+use crate::types::COSESign;
 use crate::Error;
 
 pub trait Serializable {
@@ -83,5 +86,37 @@ where
         W: std::io::Write,
     {
         ciborium::ser::into_writer(self, &mut writer).map_err(Error::from)
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct StoredItem {
+    pub public_key: PublicKey,
+    pub to1d: COSESign,
+}
+
+impl Serializable for StoredItem {
+    fn deserialize_from_reader<R>(reader: R) -> Result<Self, Error>
+    where
+        R: std::io::Read,
+    {
+        let contents: ParsedArray<ParsedArraySize2> = ParsedArray::deserialize_from_reader(reader)?;
+
+        let public_key = contents.get(0)?;
+        let to1d = contents.get(1)?;
+
+        Ok(StoredItem { public_key, to1d })
+    }
+
+    fn serialize_to_writer<W>(&self, writer: W) -> Result<(), Error>
+    where
+        W: std::io::Write,
+    {
+        let mut contents: ParsedArrayBuilder<ParsedArraySize2> = ParsedArrayBuilder::new();
+        contents.set(0, &self.public_key)?;
+        contents.set(1, &self.to1d)?;
+        let contents = contents.build();
+
+        contents.serialize_to_writer(writer)
     }
 }
